# services/platform/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tale Platform API",
    "version": "1.2.0",
...
  },
  "servers": [
    {
      "url": "",
      "description": "Same origin"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/knowledge-entries": {
      "get": {
        "tags": ["Knowledge"],
        "summary": "List knowledge entries",
        "description": "Newest first; `cursor` is the previous page’s `continueCursor`.",
        "operationId": "listKnowledgeEntries",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "The previous page’s `continueCursor` — opaque; omit for the first page"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Page size, 1..100 (default 25; out-of-range values are clamped)"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "`active` (default) or `superseded`"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["page", "isDone", "continueCursor"],
                  "properties": {
                    "page": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/KnowledgeEntry"
                      }
                    },
                    "isDone": {
                      "type": "boolean"
                    },
                    "continueCursor": {
                      "type": "string",
                      "description": "Pass back as `cursor` for the next page; empty when `isDone`"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Knowledge"],
        "summary": "Create a knowledge entry",
        "description": "Charged against the per-organization `knowledge:mutate` budget the in-app editor shares.",
        "operationId": "createKnowledgeEntry",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["topic", "content"],
                "properties": {
                  "topic": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "content": {
                    "type": "string",
                    "maxLength": 100000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created — the entry’s id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id"],
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "An active entry with this topic exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-entries/{id}": {
      "get": {
        "tags": ["Knowledge"],
        "summary": "Get a knowledge entry",
        "operationId": "getKnowledgeEntry",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Entry ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntry"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Entry not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Knowledge"],
        "summary": "Supersede a knowledge entry",
        "description": "Entries are immutable: an update writes a NEW row and answers its id; the old row becomes `superseded`. Both `topic` and `content` are required.",
        "operationId": "updateKnowledgeEntry",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Entry ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["topic", "content"],
                "properties": {
                  "topic": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "content": {
                    "type": "string",
                    "maxLength": 100000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The NEW row",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id"],
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Entry not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Entry is not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Knowledge"],
        "summary": "Delete a knowledge entry",
        "operationId": "deleteKnowledgeEntry",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Entry ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Entry not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An organization API key. Create one in Settings → API."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong"
          },
          "code": {
            "type": "string",
            "description": "A stable refusal code, when the domain has one"
          }
        }
      },
...
      "KnowledgeEntry": {
        "type": "object",
        "required": [
          "id",
          "topic",
          "content",
          "status",
          "source",
          "createdBy",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "topic": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["active", "superseded"]
          },
          "source": {
            "type": "string"
          },
          "documentId": {
            "type": "string"
          },
          "supersededBy": {
            "type": "string",
            "description": "The row that replaced this one"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "number",
            "description": "Epoch ms"
          }
        }
      }
    }
  }
}